exact solution with modes corrected

+++ src/exact_solutions.py
@@ -14,7 +14,7 @@ def GreenFunctionImages(k, H, XY, x_0, y_0, M = 100):
 def GreenFunctionModes(k, H, XY, x_0, y_0, M = 20):
     n = np.arange(0,M+1)
     beta_n = np.emath.sqrt(k**2 - (n*np.pi/H)**2)
-    norm = np.full_like(n,1/H)
+    norm = np.full_like(n,1/H, dtype=np.float64)
     norm[0] = 1/2*H
     G = -np.sum( norm*np.exp(1j*np.outer(np.abs(XY[:,0] - x_0),beta_n))/(2*1j*beta_n)*np.cos(np.pi*np.outer(XY[:,1],n)/H)*np.cos(n*np.pi*y_0/H), -1)
     return G
